reupload with fixes
laptop that this proj is on isnt linked to github so its hard to get good quality pushes but i did an overhaul of chording and added PID binding since the overlay jumping into random selected windows was annoying, also had problems on a different desktop with streamproofing so i've added checks and tried to improve the logic behind it. ctrl+q clicking a tile has been removed as it was detected if used improperly and is overall useless for a streamable overlay anyways. (also added detection for flags to help with chording efficiency)